chore(refactor): extra pathValidationUtils from renderDataUtils
